requirements/prod.txt
boto3==1.24.12
botocore==1.27.12
django-cors-headers==3.13.0
django-heroku==0.3.1
django-storages==1.12.3
gunicorn==20.1.0
python-decouple==3.6